requirements.txt
openpyxl==3.1.2
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
PuLP==2.8.0
streamlit==1.34.0
xlrd==2.0.1
